emergent-physics-comm/requirements.txt
torch>=2.0
torchvision>=0.15
numpy>=1.24
scipy>=1.10
matplotlib>=3.7
pillow>=9.0
pyyaml>=6.0
transformers>=4.30
pybullet>=3.2
kubric>=0.3
